fixed formatting on right side of page
The space after "GM: Create a BLANK adventure" was acting strangely, and after some experimenting it looks like this is because of how bootstrap handles text in and out of the container div. I moved the header outside the container and put the paragraph in a row inside the container and that appears to have solved things. This is the only place this happens, so I think we're ready for at least a small launch.